--- src/vs/workbench/browser/positronDataGrid/components/dataGridScrollbar.css
@@ -1,53 +1,0 @@
-/*---------------------------------------------------------------------------------------------
- *  Copyright (C) 2023-2024 Posit Software, PBC. All rights reserved.
- *  Licensed under the Elastic License 2.0. See LICENSE.txt for license information.
- *--------------------------------------------------------------------------------------------*/
-
-.data-grid
-.data-grid-scrollbar {
-	/* The Monaco sash is at a z-index of 35, so go lower than that. */
-	z-index: 26;
-	/* contain: strict; */
-	position: absolute;
-	box-sizing: border-box;
-	/* Makes some devices run their hardware acceleration. */
-	/* transform: translate3d(0px, 0px, 0px); */
-	background-color: var(--vscode-positronDataGrid-background);
-}
-
-.data-grid
-.data-grid-scrollbar.vertical {
-	top: -1px;
-	right: 0;
-	border-top: 1px solid var(--vscode-positronDataGrid-border);
-	border-left: 1px solid var(--vscode-positronDataGrid-border);
-}
-
-.data-grid
-.data-grid-scrollbar.horizontal {
-	left: -1px;
-	bottom: 0;
-	border-top: 1px solid var(--vscode-positronDataGrid-border);
-	border-left: 1px solid var(--vscode-positronDataGrid-border);
-}
-
-.data-grid
-.data-grid-scrollbar
-.data-grid-scrollbar-slider {
-	position: absolute;
-	background-color: var(--vscode-scrollbarSlider-background);
-}
-
-.data-grid
-.data-grid-scrollbar
-.data-grid-scrollbar-slider.vertical {
-	right: 0;
-	position: absolute;
-}
-
-.data-grid
-.data-grid-scrollbar
-.data-grid-scrollbar-slider.horizontal {
-	bottom: 0;
-	position: absolute;
-}
